pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "git-wise"
version = "0.1.0"
description = "AI-powered Git commit message generator"
readme = "README.md"
authors = [{ name = "varhuman", email = "varhuman@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["git", "commit", "AI", "OpenAI"]
dependencies = [
    "click>=8.0.0",
    "rich>=10.0.0",
    "gitpython>=3.1.0",
    "openai>=0.27.0",
    "pyyaml>=6.0.0",
    "PyGithub",
    "GitPython",
]
requires-python = ">=3.7"

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

[project.scripts]
git-wise = "git_wise.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["git_wise*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
addopts = "--cov=git_wise"
testpaths = [
    "tests",
]
